eliminate long and short from the codebase
size_t for byte counts, ptrdiff_t for signed byte counts (uniform
pointer sizes that encode read-only via negative sign), uint64_t
for 8-byte-aligned uniform buffers. Keep long/unsigned only where
required by interfaces (ftell, __builtin, stb).

+++ slides/gradient.c
@@ -10,7 +10,7 @@ typedef struct {
 static void grad_2stop_draw(slide *s, int w, int h, int y0, int y1, void *buf,
                                umbra_draw_layout const *lay, struct umbra_program *program) {
     int       uni_len = lay->uni_len;
-    long long uni_[8] = {0};
+    uint64_t  uni_[8] = {0};
     char     *uni = (char *)uni_;
     slide_uni_f32(uni, lay->shader, s->grad, 3);
     slide_uni_f32(uni, lay->shader + 12, s->color, 8);
@@ -29,10 +29,10 @@ static void grad_lut_draw(slide *s, int w, int h, int y0, int y1, void *buf,
                              umbra_draw_layout const *lay, struct umbra_program *program) {
     grad_lut_state *st = s->state;
     int             uni_len = lay->uni_len;
-    long long       uni_[8] = {0};
+    uint64_t        uni_[8] = {0};
     char           *uni = (char *)uni_;
     slide_uni_f32(uni, lay->shader, s->grad, 4);
-    slide_uni_ptr(uni, (lay->shader + 16 + 7) & ~7, st->lut, (long)(st->lut_n * 4 * 4));
+    slide_uni_ptr(uni, (lay->shader + 16 + 7) & ~7, st->lut, (ptrdiff_t)(st->lut_n * 4 * 4));
     int       ps = lay->ps;
     size_t plane_sz = (size_t)w * (size_t)h * lay->pixel_bytes;
     umbra_buf ubuf[5];
